ai-review/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.26.4
scikit-learn==1.7.0
joblib==1.3.2
better-profanity==0.7.0
python-dotenv==1.0.0
